+++ projects/ngx-webrtc-lib/src/lib/services/index.ts
@@ -1,2 +1,3 @@
 export * from './dialog.service';
+export * from './video-call-dialog.service';
 export * from './web-rtc.service';